Rename ndarray::si module -> ndarray::slice
