+++ src/lib.rs
@@ -105,7 +105,7 @@ pub use dimension::NdIndex;
 pub use dimension::IxDynImpl;
 pub use indexes::{indices, indices_of};
 pub use error::{ShapeError, ErrorKind};
-pub use si::{Si, S};
+pub use slice::{Si, S};
 
 use iterators::Baseiter;
 use iterators::{ElementsBase, ElementsBaseMut, Iter, IterMut};
@@ -143,7 +143,7 @@ mod free_functions;
 pub use free_functions::*;
 pub use iterators::iter;
 
-mod si;
+mod slice;
 mod layout;
 mod indexes;
 mod iterators;
